Switching gears for a moment. I want about 14 new themes that are more bold designs that incorporate photos (from royalty-free apis) and more bold gradient-heavy designs. I like giving users as many options as possible for the design of their software. Feel free to get creative with these 14 themes and create a creative, photo and icon-driven experience in light and dark modes -- with gradients.
Finally, please add subtle animations to the platform overall. A general suspense transition screen for page loads may be to have an eye like in the logo blinking at different times as a vector animation.

Also, "EagleEyED" is a brand name, so please add "TM" to all labels with it in superscript?