I'm still getting the error below. I don't know if you actually changed any files:
Error: Attempted to call useTheme() from the server but useTheme is on the client. It's not possible to invoke a client function from the server, it can only be rendered as a Component or passed to props of a Client Component.

src/app/layout.tsx (45:36) @ ThemeSpecificBackgroundSetterWrapper

  43 | // Wrapper component to correctly use useTheme from next-themes
  44 | function ThemeSpecificBackgroundSetterWrapper() {
> 45 |   const { resolvedTheme } = useTheme(); // Use the hook from next-themes
     |                                    ^
  46 |
  47 |   React.useEffect(() => {
  48 |     const setBg = async () => {
Call Stack
6

Show 4 ignore-listed frame(s)
ThemeSpecificBackgroundSetterWrapper
src/app/layout.tsx (45:36)
RootLayout
src/app/layout.tsx (98:15)